requirements.txt
pandas
numpy
alpha_vantage
alpaca-trade-api
python-dotenv
